feat(frontend): create floodFormValidation
